requirements.txt
Flask==2.3.2
flask-cors==4.0.0
scikit-learn==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==21.2.0
